Add files via upload
1. Watch and listen to the presentation on sorting and on reading in strings.
If you have any questions, make sure that you ask one of the Demonstrators after the presentation.
2. Follow the instructions to create an application p3q2 to use Insertion Sort to sort an array of
integers.
Ensure that you know how this code works. Experiment with changing various constants and
other pieces of code.
Four files are available for testing purposes; these are a file of 10 000 unsorted integers, a file
with the 10 000 integers in sorted order, a file of 100 000 unsorted integers and a file with
the 100 000 integers in sorted order, and they are available on the Brightspace page in the
files rand-10000.txt, sorted-10000.txt, rand-100000.txt and sorted-100000.txt,
respectively.
3. Modify your answer for Question 2 to avoid unnecessary swapping of elements when inserting
an element in its correct position. Run this on the four files. Save this application as p3q3.
4. Follow the instructions to create an application p3q4 to to use Quicksort to sort an array of
integers. Run this on the four file of integers. What do you observe about the performance
of this algorithm on the four files?
5. Modify your answer for Question 4 to choose the median of the first element of the sequence,
the last element and the middle element as the pivot when partitioning the sequence. This
is known as the “Median of three” strategy. Save this application as p3q5. Run this on the
four files of integers. What do you observe about the performance of this algorithm?
6. Write an application to read in the file indie-1990s-2000s.txt (available on the Brightspace
page) and store it in a two-dimensional array for sorting. Save this application as p3q6.